Restructuring kpi-queries and unit test fix
